Functions/chicago-taxi-trip-etl/requirements.txt
functions-framework==3.*
google-cloud-dataproc